# Improve Array.copy performance