--- a/test/benchmarks/src/main/scala/scala/collection/mutable/ArrayBufferBenchmark.scala
+++ b/test/benchmarks/src/main/scala/scala/collection/mutable/ArrayBufferBenchmark.scala
@@ -45,6 +45,17 @@ class ArrayBufferBenchmark {
     bh.consume(b)
   }
 
+  @Benchmark def toObjArrayTagged(bh:Blackhole):Unit = {
+    val res = ref.asInstanceOf[ArrayBuffer[Integer]].toArray
+    bh.consume(res)
+  }
+
+  @Benchmark def toObjArrayUntagged(bh:Blackhole):Unit = {
+    val res = ref.asInstanceOf[ArrayBuffer[AnyRef]].toArray
+    bh.consume(res)
+  }
+
+
   @Benchmark def update(bh: Blackhole): Unit = {
     val b = ref.clone()
     var i = 0