@SethTisue 2024-12-20
This looks quite appealing, thank you! We'll need a signed CLA (see the "checks" section). And we might not look at this until January.

@scala/collections