@mpollmeier 2024-12-20
Big shoutout to yourkit profiler btw, this issue was immediately obvious in their flamegraph view. They hand out free licenses to open source projects, and this codepath was actually triggered by one of our open source projects (https://github.com/joernio/flatgraph/blob/066654dd286293b8e80f1a75e3fb8c147531220f/core/src/main/scala/flatgraph/DiffGraphApplier.scala#L444)