@SethTisue 2025-01-07
rebased on top of #10978 — hopefully that'll fix CI