@som-snytt 2025-01-08
I commented on the linked PR that for just building arrays using `addOne`, `ArrayBuilder.ofRef` seems to avoid the gotcha. (untested)

Also worth noting that although some workflows require "all conversations must be marked resolved", that is not the case (so far) in this repo.

Someday github will collect all the humor and bury it in the Comedy Vault under the North Pole, for future generations and alien archaeologists.