@sjrd 2025-01-16
> I couldn't (easily) figure out how `(ref: ArrayBuffer[Int]).asInstanceOf[ArrayBuffer[Integer]].toArray` performance is affected by this PR, could someone explain?

`ArrayBuffer` always uses an `Array[Object]` inside. Calling `toArray` when it's typed as `ArrayBuffer[Integer]` thus instantiates an `Array[Integer]`. It then copies its `Array[Object]` into the `Array[Integer]`. This combination previously used the slow path, and now takes the fast `System.arraycopy` path.