@lrytz 2025-01-16
Thanks! I also missed the `copyToArray` override in `ArrayBuffer`.